refactor: rename utils.rs to trace.rs
